BAH-4721 | Refactor. Move FormattedMedicationRequest To Widget Layer And Use FHIR Priority For STAT
