+++ packages/bahmni-widgets/src/medications/__tests__/__mocks__/medicationMocks.ts
@@ -1,3 +1,4 @@
+import { MedicationRequest, MedicationStatus } from '@bahmni/services';
 import { MedicationRequest as FhirMedicationRequest } from 'fhir/r4';
 
 export const fhirMedicationRequestMock: FhirMedicationRequest = {
@@ -10,4 +11,153 @@ export const fhirMedicationRequestMock: FhirMedicationRequest = {
     reference: 'Medication/test-med',
     display: 'Paracetamol',
   },
+};
+
+export const mockMedications: MedicationRequest[] = [
+  {
+    id: '1',
+    name: 'Paracetamol 500mg',
+    dose: { value: 500, unit: 'mg' },
+    quantity: { value: 30, unit: 'tablets' },
+    startDate: '2024-01-15T10:00:00Z',
+    orderDate: '2024-01-15T09:00:00Z',
+    orderedBy: 'Dr. Smith',
+    status: MedicationStatus.Active,
+    isImmediate: false,
+    asNeeded: false,
+    priority: 'routine',
+    instructions: 'Take with food',
+    fhirResource: fhirMedicationRequestMock,
+  },
+  {
+    id: '2',
+    name: 'Aspirin 100mg',
+    dose: { value: 100, unit: 'mg' },
+    quantity: { value: 14, unit: 'tablets' },
+    startDate: '2024-01-10T08:00:00Z',
+    orderDate: '2024-01-10T07:30:00Z',
+    orderedBy: 'Dr. Johnson',
+    status: MedicationStatus.OnHold,
+    isImmediate: true,
+    asNeeded: false,
+    priority: 'stat',
+    instructions: 'After meals',
+    fhirResource: fhirMedicationRequestMock,
+  },
+  {
+    id: '3',
+    name: 'Ibuprofen 200mg',
+    dose: { value: 200, unit: 'mg' },
+    quantity: { value: 20, unit: 'tablets' },
+    startDate: '2024-01-08T14:00:00Z',
+    orderDate: '2024-01-08T13:45:00Z',
+    orderedBy: '',
+    status: MedicationStatus.Completed,
+    isImmediate: false,
+    asNeeded: true,
+    priority: 'routine',
+    instructions: 'As needed for pain',
+    fhirResource: fhirMedicationRequestMock,
+  },
+];
+
+export const mockMedicationWithDoseForm: MedicationRequest = {
+  id: '4',
+  name: 'Paracetamol 500 mg',
+  dose: { value: 500, unit: 'mg' },
+  quantity: { value: 10, unit: 'Tablets' },
+  startDate: '2024-01-15T10:00:00Z',
+  orderDate: '2024-01-15T09:00:00Z',
+  orderedBy: 'Dr. Smith',
+  status: MedicationStatus.Active,
+  isImmediate: false,
+  asNeeded: false,
+  priority: 'routine',
+  instructions: 'Take with food',
+  doseForm: 'Tablet',
+  fhirResource: fhirMedicationRequestMock,
+};
+
+export const mockMedicationWithoutDoseForm: MedicationRequest = {
+  id: '5',
+  name: 'Aspirin 100mg',
+  dose: { value: 100, unit: 'mg' },
+  quantity: { value: 14, unit: 'tablets' },
+  startDate: '2024-01-10T08:00:00Z',
+  orderDate: '2024-01-10T07:30:00Z',
+  orderedBy: 'Dr. Johnson',
+  status: MedicationStatus.Active,
+  isImmediate: false,
+  asNeeded: false,
+  priority: 'routine',
+  instructions: 'After meals',
+  fhirResource: fhirMedicationRequestMock,
+};
+
+export const mockMedicationCapsule: MedicationRequest = {
+  id: '6',
+  name: 'Vitamin A 50000 IU',
+  dose: { value: 50000, unit: 'IU' },
+  quantity: { value: 1, unit: 'Capsule' },
+  startDate: '2024-01-15T10:00:00Z',
+  orderDate: '2024-01-15T09:00:00Z',
+  orderedBy: 'Dr. Smith',
+  status: MedicationStatus.Active,
+  isImmediate: true,
+  asNeeded: false,
+  priority: 'stat',
+  instructions: 'Once',
+  doseForm: 'Capsule',
+  fhirResource: fhirMedicationRequestMock,
+};
+
+export const mockMixedDoseFormMedications: MedicationRequest[] = [
+  {
+    id: '7',
+    name: 'Paracetamol',
+    dose: { value: 500, unit: 'mg' },
+    quantity: { value: 2, unit: 'Tablet' },
+    startDate: '2024-01-15T10:00:00Z',
+    orderDate: '2024-01-15T09:00:00Z',
+    orderedBy: 'Dr. Smith',
+    status: MedicationStatus.Active,
+    isImmediate: false,
+    asNeeded: false,
+    priority: 'routine',
+    instructions: 'TID',
+    doseForm: 'Tablet',
+    fhirResource: fhirMedicationRequestMock,
+  },
+  {
+    id: '8',
+    name: 'Custom Liquid',
+    dose: { value: 5, unit: 'ml' },
+    quantity: { value: 1, unit: 'bottle' },
+    startDate: '2024-01-15T10:00:00Z',
+    orderDate: '2024-01-15T09:00:00Z',
+    orderedBy: 'Dr. Smith',
+    status: MedicationStatus.Active,
+    isImmediate: false,
+    asNeeded: false,
+    priority: 'routine',
+    instructions: 'BID',
+    fhirResource: fhirMedicationRequestMock,
+  },
+];
+
+export const mockStatMedication: MedicationRequest = {
+  id: '9',
+  name: 'Urgent Drug',
+  dose: { value: 100, unit: 'mg' },
+  quantity: { value: 1, unit: 'vial' },
+  startDate: '2024-01-15T10:00:00Z',
+  orderDate: '2024-01-15T09:00:00Z',
+  orderedBy: 'Dr. Smith',
+  status: MedicationStatus.Active,
+  isImmediate: true,
+  asNeeded: false,
+  priority: 'stat',
+  instructions: 'Once',
+  doseForm: 'IV Injection',
+  fhirResource: fhirMedicationRequestMock,
 };
