+++ packages/bahmni-widgets/src/medications/constants.ts
@@ -1,0 +1,6 @@
+export enum MEDICATION_REQUEST_PRIORITY {
+  STAT = 'stat',
+  ROUTINE = 'routine',
+  URGENT = 'urgent',
+  ASAP = 'asap',
+}
